Refactor maintenance into own module and add UpdateMaintenance
